Time: 5 ms (21.74%), Space: 34.1 MB (35.68%) - LeetHub

+++ 0202-happy-number/0202-happy-number.kt
@@ -1,11 +1,11 @@
 class Solution {
     fun isHappy(n: Int): Boolean {
-        var newN = n.toLong()
-        val map = HashMap<Long, Boolean>()
+        var newN = n.toInt()
+        val map = HashMap<Int, Boolean>()
         
-        while (newN > 1L) {
+        while (newN > 1) {
             newN = newN.toString().sumOf { 
-                it.digitToInt() * it.digitToInt().toLong()
+                it.digitToInt() * it.digitToInt()
             }
             
             if (map[newN] == null) {
